% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundConstr.bs.R, R/boundConstr.ns.R,
%   R/boundConstr.ps.R
\name{boundConstr.bs}
\alias{boundConstr.bs}
\alias{boundConstr.ns}
\alias{boundConstr.ps}
\title{Boundary constraints for spline bases}
\usage{
\method{boundConstr}{bs}(x, value = 0, side = "right", thr = NULL,
  sm = NULL, ...)

\method{boundConstr}{ns}(x, ...)

\method{boundConstr}{ps}(x, ...)
}
\arguments{
\item{x}{A B-spline basis object.}

\item{value}{A numerical indicating the boundary constraint value. Default to 0.}

\item{side}{Character indicating the side on which the constraint applies. One of \code{"right"} (the default), \code{"left"} or \code{"both"}.}

\item{thr}{A numeric value indicating a threshold from which the smooth should be equal to \code{value}. Default to the boundary knots of the spline basis. Cannot be used with \code{side = both}. See details.}

\item{sm}{A positive integer controlling the smoothness of the constraint. Ignored when \code{thr} is provided. See details.}

\item{...}{Additional parameters passed to or from other methods.}
}
\value{
A list containing the constraint matrix \code{Cmat}, and lower/upper bound vectors (\code{lb} and \code{ub}, respectively).
}
\description{
Method for constraining a smooth functions estimated using B-Spline bases to reach a given value at its boundary.
}
\details{
See the \link[=boundConstr]{generic} method for an overview of bound constraints.
\subsection{Boundary constraints with B-splines}{

By construction, B-spline basis functions sum to one at any point. Therefore, boundary constraints can essentially be enforced through equality constraints on the first (if \code{side = left}) or last (if \code{side = right}) few coefficients of the B-spline basis, with small variations for \link[splines:ns]{natural} or \link[dlnm:ps]{penalised} splines.
}

\subsection{The \code{thr} and \code{sm} arguments}{

The \code{sm} argument can control how smoothly the resulting function reaches the constraint \code{value}. If only one basis is non-null at the boundary and need its coefficient to be constrained for the boundary constraint, the number of bases that reach zero exactly at the boundary is equal to the degree of the spline. Forcing their coefficients to equal \code{value} results in a smoother convergence at the boundary. Note that increasing \code{sm} to be higher than the degree of the spline will propagate the equality constraint towards interior knots.

Alternatively, the user can provide a threshold through \code{thr} above which (if \code{side = right}) or below which (if \code{side = left}) the resulting function is guaranteed to be equal to \code{value}. This allows propagating the boundary constraint towards the interior of the range. Note that where the constraint starts depends on the knots, as using \code{thr} will actually enforce the constraint from the maximum (minimum) interior knot such as \code{thr} is included in the constrained range. This does not allow \code{side = both} so if thresholds are needed on both sides of the range, it needs to be done with two calls to \code{boundConstr}.
}
}
\examples{
###############################################################################
# Bound constraints for bs: warming dataset

library(splines)

# Force the fit to reach the last value
# Remove the intercept, so that it is included in the constraint
basis <- bs(warming$year, df = 10)
mod <- glm(anomaly ~ 0 + basis, data = warming, method = "cirls.fit",
  constr = ~ bound(basis, value = 0.75))

# Plot result
plot(anomaly ~ year, data = warming, xlab = "", ylab = "Temperature anomaly")
lines(warming$year, predict(mod), col = 3, lwd = 2)
}
\seealso{
The generic \link{boundConstr} functions.
}
