% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundConstr.onebasis.R, R/boundConstr.strata.R
\name{boundConstr.onebasis}
\alias{boundConstr.onebasis}
\alias{boundConstr.strata}
\title{Boundary constraints for onebasis objects}
\usage{
\method{boundConstr}{onebasis}(x, ...)

\method{boundConstr}{strata}(x, value = 0, side = "right", thr = NULL,
  sm = 1, ...)
}
\arguments{
\item{x}{A \code{onebasis} object or any other objects from the \code{dlnm} package.}

\item{...}{Additional parameters passed to or from other methods.}

\item{value}{A numerical indicating the boundary constraint value. Default to 0.}

\item{side}{Character indicating the side on which the constraint applies. One of \code{"right"} (the default), \code{"left"} or \code{"both"}.}

\item{thr}{An integer indicating a threshold from which the association should be equal to \code{value}. Using \code{thr} overrides \code{sm}. Cannot be used with \code{side = both}.}

\item{sm}{A positive integer indicating the degree of smoothness for the constraint. Typically indicates the number of coefficients to constrain to \code{value}.}
}
\value{
A list containing the constraint matrix \code{Cmat}, and lower/upper bound vectors (\code{lb} and \code{ub}, respectively).
}
\description{
Method to set up boundary constraints for basis objects constructed through \link[dlnm:onebasis]{onebasis} and other specifications from the \link[dlnm:dlnm]{dlnm} package.
}
\details{
The \code{onebasis} method simply check there is an available \link{boundConstr} method available for the underlying basis type (for instance \link[=boundConstr.strata]{strata} or \link[=boundConstr.ps]{P-splines}) and dispatches \code{x} to the appropriate method. The ellipsis \code{...} is used to pass the necessary arguments to the method, including \code{value} and \code{side}. See the help for the specific methods.

The \code{strata} method allows to specify boundary constraints for \link[dlnm:strata]{strata} objects, which are factors expanded through dummy parametrisation. In this context, boundary constraints essentially add an equality constraint to \code{value} on the \code{sm} extreme coefficients (depending on \code{side}). A numeric threshold \code{thr} can be specified instead of \code{sm}, which will add the equality constraint from the strata that contains \code{thr}. This allows to propagate boundary constraints towards the interior of the range.
}
\examples{
###############################################################################
# Bound constraints for onebasis: warming dataset

library(dlnm)

# Force the fit to reach the last value
# Remove the intercept, so that it is included in the constraint
basis <- onebasis(warming$year, fun = "strata", df = 10, intercept = TRUE)
mod <- glm(anomaly ~ 0 + basis, data = warming, method = "cirls.fit",
  constr = ~ bound(basis, value = 0.75))

# Plot result
plot(anomaly ~ year, data = warming, xlab = "", ylab = "Temperature anomaly")
lines(warming$year, predict(mod), col = 2, lwd = 2)
}
\seealso{
The generic \link{boundConstr} functions. The \link[=boundConstr.crossbasis]{crossbasis} is used to specify the constraint on distributed lag (non-linear) models.
}
