% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundConstr.R, R/boundConstr.default.R,
%   R/boundConstr.factor.R, R/boundConstr.bs.R, R/boundConstr.ns.R,
%   R/boundConstr.ps.R, R/boundConstr.strata.R, R/boundConstr.onebasis.R
\name{boundConstr}
\alias{boundConstr}
\alias{bound}
\alias{boundConstr.default}
\alias{boundConstr.factor}
\alias{boundConstr.bs}
\alias{boundConstr.ns}
\alias{boundConstr.ps}
\alias{boundConstr.strata}
\alias{boundConstr.onebasis}
\title{Boundary constraints}
\usage{
boundConstr(x, ...)

\method{boundConstr}{default}(x, value = 0, side = c("right", "left",
  "both"), deg = NULL, intercept = FALSE, ...)

\method{boundConstr}{factor}(x, intercept = FALSE, ...)

\method{boundConstr}{bs}(x, deg = NULL, ...)

\method{boundConstr}{ns}(x, ...)

\method{boundConstr}{ps}(x, ...)

\method{boundConstr}{strata}(x, ...)

\method{boundConstr}{onebasis}(x, ...)
}
\arguments{
\item{x}{An object representing a design matrix of predictor variables, typically basis functions. See Details for supported objects.}

\item{...}{Additional parameters passed to or from other methods. Includes \code{value} and \code{deg} for most methods.}

\item{value}{A numerical indicating the boundary constraint value. Default to 0.}

\item{side}{Character indicating the side on which the constraint applies. One of \code{"right"} (the default), \code{"left"} or \code{"both"}.}

\item{deg}{A positive integer indicating the degree of smoothness for the constraint. The default value is different according to the method. See details.}

\item{intercept}{For the default and factor method, a logical value indicating if the design matrix includes an intercept. In most cases, it will be automatically extracted from \code{x}, but this argument can be used to override it.}
}
\value{
A list containing the constraint matrix \code{Cmat}, and lower/upper bound vectors (\code{lb} and \code{ub}, respectively).
}
\description{
Builds a constraint matrix and bound vectors for constraining the boundary of splines and other nonlinear functions to a given value. This is a generic function designed to be used in the \link[=buildCmat]{constr} formula interface. It allows methods for a wide range of regression terms.
}
\details{
Enforcing boundary constraints amounts to imposing an equality constraint on the \code{deg} first (if \code{side = left}), last (if \code{side = right}), or both (if \code{side = both}) coefficients of the basis matrix \code{x}. The equality constraint sets the bounds to \code{lb = ub = value}.
\subsection{Usage}{

The recommended usage is to use this function through a call to \code{bound} on a term in the \link[=buildCmat]{constr} formula interface. This method is then called internally to create the constraint matrix and bound vectors. However, \code{boundConstr} can also be called directly on a matrix-like object to manually build or inspect the constraint matrix.

All methods internally rely on the default method for general matrices. Unless specified, all methods use the same parameters as the default one, which are passed through \code{...}. The only exception is \code{intercept} which is often inferred from the \code{x} object itself. In a typical usage in which \code{boundConstr} is called from the \code{constr} argument, \code{intercept} is automatically determined from the \link[stats:glm]{glm} formula.
}

\subsection{The \code{deg} parameter and spline bases}{

The \code{deg} parameter indicates the number of coefficients on the left/right that are constrained to be equal to \code{value} and can be interpreted as a smoothness degree for the boundary constraint. The default is different for each method. It is set to 1 for the default method or the methods related to categorical variables (e.g. \code{factor} and \code{strata}).

For B-spline bases methods (such as \code{bs}, \code{ps} and \code{ns}), the default is to be equal to the degree of the spline. This corresponds to the number of bases that are non-null at the boundary, are therefore that contribute to the value of the smooth at the boundary. Since, by construction, B-spline basis functions sum to one at any point, constraining all coefficients to \code{value} will result in the smooth being equal to \code{value} at the boundary. Note that \code{deg} can be reduced for \code{bs} and \code{ps} (but not \code{ns}),  resulting in a less smooth convergence towards \code{value}.
}

\subsection{Available methods}{

In addition to the default method, \code{boundConstr} currently supports methods for several classes. The full list can also be consulted through \code{methods(shapeConstr)}.
\subsection{Categorical variables}{
\itemize{
\item \link{factor}: for \code{factor} objects. Extract the \link[stats:contrasts]{contrasts} to define the constraint matrix. Here the \code{intercept} argument has the same interpretation as in the default method, i.e. if set to \code{TRUE} it means the \code{glm} model does not include an intercept externally to the factor. Note that, in this case, a simple dummy coding is done in R.
\item \link[dlnm:strata]{strata}: Indicator variables defining strata from the \link[dlnm:dlnm]{dlnm} package. Here the shape is applied to the coefficient of strata, considering strata like a categorical variable.
}
}

\subsection{B-spline bases}{
\itemize{
\item \link[splines:bs]{bs}: B-splines bases from the \link{splines} package. Here \code{deg} is set by default to the \code{degree} of the basis, but can be reduced for a less smooth constraint.
\item \link[splines:ns]{ns}: B-spline bases for natural cubic splines from the \link{splines} package. Builds constraints for \code{bs} that are then adjusted for \code{ns} specifically.
\item \link[dlnm:ps]{ps}: P-spline bases from the \link[dlnm:dlnm]{dlnm} package. For \code{ps}, \code{deg} should not be reduced to a lower value than its degree.
}
}

\subsection{Distributed-lag linear and nonlinear models (DLNM)}{
\itemize{
\item \link[dlnm:onebasis]{onebasis}: General method for basis functions generated in the package. Internally calls the relevant method for the specified basis.
\item \link[=boundConstr.crossbasis]{crossbasis}: Most exhaustive method to constrain DLNMs. See its dedicated \link[=boundConstr.crossbasis]{help page}.
}
}

}
}
\examples{
###############################################################################
# Bound constraints example: warming dataset

library(splines)

# Force the fit to reach the last value
# Remove the intercept, so that it is included in the constraint
model <- glm(anomaly ~ 0 + decade, data = warming, method = "cirls.fit",
  constr = ~ bound(decade, value = 0.75))

# Same but with splines
basis <- ns(warming$year, df = 10)
splmodel <- glm(anomaly ~ 0 + basis, data = warming, method = "cirls.fit",
  constr = ~ bound(basis, value = 0.75))

# Plot result
plot(anomaly ~ year, data = warming, xlab = "", ylab = "Temperature anomaly")
lines(warming$year, predict(model), col = 2, lwd = 2)
lines(warming$year, predict(splmodel), col = 3, lwd = 2)
}
\seealso{
\link{buildCmat} detailing the \code{constr} interface.
}
