% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCmat.R
\name{checkCmat}
\alias{checkCmat}
\title{Check constraint matrix irreducibility}
\usage{
checkCmat(Cmat)
}
\arguments{
\item{Cmat}{A constraint matrix as passed to \code{\link[=cirls.fit]{cirls.fit()}}}
}
\value{
A list with two elements:
\item{redundant}{Vector of indices of redundant constraints}
\item{equality}{Indicates which constraints are part of an underlying equality constraint}
}
\description{
Checks a constraint matrix does not contains redundant rows
}
\details{
The user typically doesn't need to use \code{checkCmat} as it is internally called by \code{\link[=cirls.control]{cirls.control()}}. However, it might be useful to undertsand if \code{Cmat} can be reduced for inference purpose. See the note in \code{\link[=confint.cirls]{confint.cirls()}}.

A constraint matrix is irreducible if no row can be expressed as a \emph{positive} linear combination of the other rows. When it happens, it means the constraint is actually implicitly included in other constraints in the matrix and can be dropped. Note that this a less restrictive condition than the constraint matrix having full row rank (see some examples).

The function starts by checking if some constraints are redundant and, if so, checks if they underline equality constraints. In the latter case, the constraint matrix can be reduced by expressing these constraints as a single equality constraint with identical lower and upper bounds (see \code{\link[=cirls.fit]{cirls.fit()}}).

The function also checks whether there are "zero constraints" i.e. constraints with only zeros in \code{Cmat} in which case they will be labelled as redundant.
}
\references{
Meyer, M.C., 1999. An extension of the mixed primal–dual bases algorithm to the case of more constraints than dimensions. \emph{Journal of Statistical Planning and Inference} \strong{81}, 13–31. \doi{10.1016/S0378-3758(99)00025-7}
}
\seealso{
\code{\link[=confint.cirls]{confint.cirls()}}
}
