% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.cirls.R, R/vcov.cirls.R
\name{confint.cirls}
\alias{confint.cirls}
\alias{vcov.cirls}
\title{Calculate Confidence Intervals and Variance-Covariance Matrix for a \code{cirls} object.}
\usage{
\method{confint}{cirls}(object, parm, level = 0.95, nsim = 1000, ...)

\method{vcov}{cirls}(object, nsim = 1000, ...)
}
\arguments{
\item{object}{A fitted \code{cirls} object.}

\item{parm}{A specification of which parameters to compute the confidence intervals for. Either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{The confidence level required.}

\item{nsim}{The number of simulations to consider. Corresponds to \code{n} in \link[TruncatedNormal:tmvnorm]{rtmvnorm}. See details().}

\item{...}{Further arguments passed to or from other methods. Currently ignored.}
}
\value{
For \code{confint}, a two-column matrix with columns giving lower and upper confidence limits for each parameter.

For \code{vcov}, a matrix of the estimated covariances between the parameter estimates of the model.
}
\description{
\code{confint.cirls} computes confidence intervals for one of more parameters in a GLM fitted via \link[=cirls.fit]{cirls.fit}. \code{vcov.cirls} compute the variance-covariance matrix of the parameters. These functions supersede the default \link[stats:confint]{confint} and \link[stats:vcov]{vcov} methods for \code{cirls} objects.
}
\details{
These functions are custom methods for \link[=cirls.fit]{cirls} objects to supersede the default methods used for \link[stats:glm]{glm} objects.

Both methods simulate \code{nsim} realisations from a Truncated Normal distribution in which the truncation is defined by the constraints. Variance-Covariance matrix and Confidence Intervals are then computed as the empirical variance-covariance matrix and quantiles, respectively.
}
\references{
Geweke, J.F., 1996. Bayesian Inference for Linear Models Subject to Linear Inequality Constraints, in: Lee, J.C., Johnson, W.O., Zellner, A. (Eds.), Modelling and Prediction Honoring Seymour Geisser. \emph{Springer, New York, NY}, pp. 248–263. \href{https://doi.org/10.1007/978-1-4612-2414-3_15}{10.1007/978-1-4612-2414-3_15}

Botev, Z.I., 2017, The normal law under linear restrictions: simulation and estimation via minimax tilting, \emph{Journal of the Royal Statistical Society, Series B}, \strong{79} (\strong{1}), pp. 1–24.
}
