% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edf.R
\name{edf}
\alias{edf}
\title{Expected degrees of freedom}
\usage{
edf(object, nsim = 10000, seed = NULL)
}
\arguments{
\item{object}{A \code{cirls} object or any object inheriting from \code{lm}. See details.}

\item{nsim}{The number of simulations.}

\item{seed}{An optional seed for the random number generator. See \link{set.seed}.}
}
\value{
A vector of length three containing the three types of degrees of freedom:
\item{udf}{The \emph{unconstrained} degrees of freedom, i.e. the rank plus any dispersion parameter for \code{glm} objects.}
\item{odf}{The \emph{observed} degrees of freedom, that is \code{udf} minus the number of active constraints.}
\item{edf}{The \emph{expected} degrees of freedom estimated by simulation, as described in the Details section.}
For \code{cirls} objects, the vector includes the simulated distribution of the number of active constraints as an \code{actfreq} attribute.
}
\description{
Estimates expected degrees of freedom (df) of a \link[=cirls.fit]{cirls} object through simulations.
}
\details{
Computes \strong{unconstrained}, \strong{observed}, and \strong{expected} df for \code{cirls} objects. The function also works for most objects inheriting from \code{lm} although, in this case, only the unconstrained (\code{udf}) part makes sense.
\subsection{Unconstrained df}{

\strong{Unconstrained} degrees of freedom (\code{udf}) refer to the usual degrees of freedom, i.e. the number of estimated parameters \eqn{p}. In GLMs, this corresponds to the number of coefficients, plus one degree of freedom for the dispersion parameter, when relevant.
}

\subsection{Observed df}{

\strong{Observed} degrees of freedom (\code{odf}) correspond to the constrained df of a fitted \code{cirls} model. This corresponds to \eqn{p-m_a}, where \eqn{m_a} is \emph{the number of active constraint} in the fitted \code{cirls} model. Intuitively, the more restricted the final fit, the more df are decreased.
}

\subsection{Expected df}{

\strong{Expected} degrees of freedom (\code{edf}) account for the sampling variation in the number of active constraints. A different sample might result in a fit with another number of active constraints \eqn{m_a} which means \code{odf} represents an inaccurate estimation of the reduction in degrees of freedom induced by the constraints. \code{edf} is defined as \eqn{p - \bar{m_a}} where \eqn{\bar{m_a}} is the expected number of active constraints. Here, \eqn{\bar{m_a}} is estimated by sampling from the \link[=uncons]{unconstrained} distribution of coefficients (see \link{simulCoef}).
}
}
\examples{
###############################################################################
# Monotone strata levels with the warming dataset

### Fit the model

# Non-decreasing constraint on decadal strata
model <- glm(anomaly ~ decade, data = warming, method = "cirls.fit",
  cons = ~ shape(decade, "inc"))

# Plot result
plot(anomaly ~ year, data = warming, xlab = "", ylab = "Temperature anomaly")
lines(warming$year, predict(model), col = 2, lwd = 2)

### Extract results

# Coefficients and confidence intervals
betas <- coef(model)
v <- vcov(model)
cis <- confint(model)

# Degrees of freedom: represent number of strata change
# ?edf
edf(model)

### Compare with an unconstrained model

# Refit the model without constraints
umodel <- uncons(model)

# Add result
lines(warming$year, predict(umodel), col = 3, lwd = 2, lty = 2)

}
\references{
Meyer, M.C., 2013. Semi-parametric additive constrained regression. \emph{Journal of Nonparametric Statistics} \strong{25}, \strong{715–730}. \href{https://doi.org/10.1080/10485252.2013.797577}{DOI:10.1080/10485252.2013.797577}
}
\seealso{
\link{logLik.cirls} which internally calls \code{edf} to compute degrees of freedom. \link{simulCoef} for coefficient simulation.
}
