% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgl.R
\docType{data}
\name{fgl}
\alias{fgl}
\title{Measurement of forensic glass fragments}
\format{
A data frame of 214 observations and 10 variables. Apart from \code{RI} and \code{type}, all other variables represent a ratio of the given chemical component.
\describe{
\item{RI}{Refractive index; more precisely the refractive index is 1.518xxxx.}
\item{Na}{Sodium}
\item{Mg}{Manganese}
\item{Al}{Aluminium}
\item{Si}{Silicon}
\item{K}{Potassium}
\item{Ca}{Calcium}
\item{Ba}{Barium}
\item{Fe}{Iron}
\item{type}{Class type (see Details)
}
}
\source{
Venables, W. N. and Ripley, B. D. (2002) \emph{Modern Applied Statistics with S}. Fourth edition. Springer.
}
\usage{
fgl
}
\description{
Dataset adapted from the \link[MASS:fgl]{fgl} dataset. Chemical composition of fragments of glass collected in forensic work.
}
\details{
The fragments were originally classed into seven types, one of which was absent in this dataset. The categories which occur are window float glass (\code{WinF}: 70), window non-float glass (\code{WinNF}: 76), vehicle window glass (\code{Veh}: 17), containers (\code{Con}: 13), tableware (\code{Tabl}: 9) and vehicle headlamps (\code{Head}: 29).

The source of this dataset is the MASS package. The dataset has been adapted for compositional data analysis by adding a small increment to null values and rescaling all chemical components to sum to unity.
}
\examples{
###############################################################################
# Compositional regression example with the fgl dataset

# Select the composition and use log
x <- as.matrix(fgl[,2:9])
z <- log(x)

# Fit model
model <- glm(RI ~ z, data = fgl, method = "cirls.fit",
  cons = ~ zerosum(z))

# Coefficients sum to 1
coef(model)
sum(coef(model)[-1])

# Confidence intervals
confint(model)

}
\keyword{datasets}
