% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/london.R
\docType{data}
\name{london}
\alias{london}
\title{Daily mortality, temperature and pollution data in London}
\format{
A data frame with 2922 observations and 12 variables:
\describe{
\item{date}{\code{Date} variable}
\item{death}{Counts of all cause and all-age mortality}
\item{age0_64}{Counts of mortality for age group 0-64}
\item{age65_74}{Counts of mortality for age group 65-74}
\item{age75_84}{Counts of mortality for age group 75-84}
\item{age85plus}{Counts of mortality for age group 85+}
\item{tmax}{Daily maximum temperature (in Celsius degree)}
\item{tmean}{Daily mean temperature (in Celsius degree)}
\item{tmin}{Daily minimum temperature (in Celsius degree)}
\item{pm10}{Daily mean coarse particulate matter concentration (\eqn{\mu g/m^3})}
\item{o3}{Daily mean ozone concentration (\eqn{\mu g/m^3})}
\item{co}{Daily mean carbon monoxide concentration (\eqn{\mu g/m^3})}
}
}
\source{
Office for National Statistics

British Atmospheric Data Centre
}
\usage{
london
}
\description{
This dataset contains daily time series of all-cause mortality for several age groups, temperature (maximum, mean and minimum) and air pollutants (PM10, ozone and carbon monoxide) for the period of 1996-2003.
}
\examples{
###############################################################################
# Non-negative coefficient example with the london dataset

library(splines)

### Association between CO and mortality
# Model includes spline of date and day-of-year
# Nonnegative constraint on CO
model <- glm(death ~ co + weekdays(date) + ns(date, df = 7*8),
  data = london, family = "quasipoisson",
  method = "cirls.fit", cons = ~ shape(co, "pos"))

# Coefficient and confidence interval
coef(model)[2]
confint(model)[2,]

}
\keyword{datasets}
