% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceCons.R
\name{reduceCons}
\alias{reduceCons}
\title{Reduce a set of linear constraints}
\usage{
reduceCons(Cmat, lb = NULL, ub = NULL, redundant = TRUE,
  equality = TRUE, warn = FALSE)
}
\arguments{
\item{Cmat}{A constraint matrix.}

\item{lb, ub}{Bound vectors. If provided, should be consistent with \code{Cmat}. if not, default to a vector of 0 and \code{Inf}, respectively.}

\item{redundant}{Logical switch indicating whether to check for redundant constraints.}

\item{equality}{Logical switch indicating whether to check for underlying equality constraints.}

\item{warn}{Logical indicating if the user should be warned when constraint reduction happens.}
}
\value{
A list with the following elements:
\item{Cmat/lb/ub}{The reduced set of constraints.}
\item{redundant}{Vector of indices indicating the redundant constraints that have been removed.}
\item{equality}{Vector of indices indicating the constraints that were part of underlying equality constraints.}
}
\description{
Internal function checking whether a set of \code{Cmat}, \code{lb} and \code{ub} contains redundant or underlying equality constraints, and attempting to reduce the set accordingly. The objective is to obtain the smallest possible set of constraints without affecting the original problem.
}
\details{
This function is mostly for internal used, being silently called by \link{buildCmat} and in some Constr functions used to create \code{Cmat}/\code{lb}/\code{ub}. Redundant or underlying equality constraints can result from putting together several sets of constraints together (for example non-decreasing and convex regression) and reducing the constraint matrix can be crucial to allow for inference. Therefore, if this check can be turned off when using \code{cirls} (see \link{cirls.control}), it is not recommended to do so.
\subsection{Redundant constraint}{

In a set of linear constraints \eqn{\mathbf{C}\beta \geq \mathbf{l}}, the constraint k is \strong{redundant} if it can be removed without affecting the feasible region of the regression. Redundancy can be checked by finding the minimum possible value for the constraint \eqn{\mathbf{c}_k^T \beta} subject to all other constraints, i.e. by solving the linear optimisation problem

\deqn{\begin{aligned}
\min \mathbf{c}_k^T \beta \\
s.t. \mathbf{C}_{-k}\beta \geq \mathbf{l}_{-k}
\end{aligned}}

where \eqn{\mathbf{C}_{-k}} and \eqn{l_{-k}} are the constraint matrix and bound vector to which the kth row has been removed.

If the minimum possible value is greater than or equal to \eqn{l_k}, then the corresponding constraint is redundant. This is easily extended for upper bounds, replacing the min by max in the problem above, and checking the maximum against the upper bound \eqn{u_k}.
}

\subsection{Underlying equality}{

There is an underlying equality constraint when a constraint is forced to be equal to its lower bound \eqn{l_k} by other constraints. If the maximum possible value for \eqn{\mathbf{c}_k^T \beta} subject to all other constraints is equal to \eqn{l_k}, then it means \eqn{\mathbf{c}_k^T \beta} is forced to be exactly \eqn{l_k} and is therefore an underlying equality constraint. This is checked by changing the min to max in the problem above.
}

\subsection{Algorithm}{

This function implements a simple algorithm to reduce the set of constraints. There is an initial check for row of \code{Cmat} that only contain zeros, as well as for constraints with infinite bounds only. Then it goes through every constraint with the following steps:
\enumerate{
\item Find the minimum and maximum possible values of \eqn{\mathbf{c}_k^T \beta} using the optimisation problem above.
\item If the minimum is higher than \eqn{l_k}, remove the bound, and same if the maximum is lower than \eqn{u_k}.
\item If no bound remains, this constraint is discarded.
\item If bounds remain, check for underlying equality constraint by comparing the minimum/maximum to \eqn{u_k/l_k}.
}

This algorithm is similar to the "naive" algorithm of Caron et al. (1989).
}
}
\note{
The result can depend on the ordering of \code{Cmat} because constraints are checked in the order provided. However, the results will not impact the feasible region in \code{cirls}, only what the matrix looks like.
}
\examples{
###################################################
# Simple example

# Two constraints: one >=0 and one >= 1
Cmat <- as.matrix(rep(1, 2))
lb <- 0:1

# The first one is removed because redundant
reduceCons(Cmat, lb)

###################################################
# Example of underlying equality constraint

# Contraint: Parameters sum is >= 0 and sum is <= 0
cmateq <- rbind(rep(1, 3), rep(-1, 3))

# Transformed into a single equality constraint
reduceCons(cmateq)

###################################################
# An example with shape constraints

library(dlnm)

# Constraints: successive coefficients should increase and be convex
# Intuitively, if the first two coefficients increase,
# then convexity forces the rest to increase which means there is redundancy
basis <- ps(1:10, df = 5)
cmatic <- rbind(
  shapeConstr(basis, shape = "inc")$Cmat, # Increasing
  shapeConstr(basis, shape = "cvx")$Cmat # Convex
)

# Checking indicates that some constraints are redundant
# Returns reduced matrix and a warning
reduceCons(cmatic)

# Compare without removing the redundant constraints
reduceCons(cmatic, redundant = FALSE)

# Note that this is silently done when both "inc" and "cvx" are provided
shapeConstr(basis, shape = c("inc", "cvx"))
}
\references{
Caron, R.J., McDonald, J.F., Ponic, C.M., 1989. A degenerate extreme point strategy for the classification of linear constraints as redundant or necessary. \emph{J Optim Theory Appl} \strong{62}, \strong{225–237}. \href{https://doi.org/10.1007/BF00941055}{DOI:10.1007/BF00941055}

Telgen, J., 1983. Identifying Redundant Constraints and Implicit Equalities in Systems of Linear Constraints. \emph{Management Science} \strong{29}, \strong{1209–1222}. \href{https://doi.org/10.1287/mnsc.29.10.1209}{DOI:10.1287/mnsc.29.10.1209}
}
\seealso{
\link{buildCmat} for how constraint matrices are built in \code{cirls}.
}
