% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundConstr.crossbasis.R,
%   R/shapeConstr.crossbasis.R
\name{boundConstr.crossbasis}
\alias{boundConstr.crossbasis}
\alias{shapeConstr.crossbasis}
\title{Constraints for distributed lag linear and non-linear models}
\usage{
\method{boundConstr}{crossbasis}(x, dim = NULL, overall = FALSE,
  slice = NULL, ...)

\method{shapeConstr}{crossbasis}(x, dim = NULL, overall = FALSE,
  slice = NULL, ...)
}
\arguments{
\item{x}{A \code{crossbasis} object.}

\item{dim}{Either \code{"var"} or \code{"lag"}. This is the dimension on which the constraint will be applied, constraining either the exposure-response or lag-response relationship, respectively. If \code{NULL}, the default is to apply the constraint on the \code{"var"} dimension if it is \emph{nonlinear} (i.e. has more than one degree of freedom), or to apply it to the \code{"lag"} dimension otherwise (i.e. when a \emph{linear} distributed lag model is specified).}

\item{overall}{Only when \code{dim = "var"}, logical indicating whether the constraint should be applied only on the overall cumulative association or for each specific lag.}

\item{slice}{A numeric vector of length 2 restricting the constraint on a specific range of the \emph{other} dimension, namely the lag (for \code{dim = "var"}) or exposure (\code{dim = "lag"}) space. By default, no slicing is performed. See Details.}

\item{...}{Parameters specific to the type of constraint to be applied. Includes for instance \code{shape} for \link{shapeConstr}, or \code{value} for \link{boundConstr}. See the main help page of the relevant generic method.}
}
\value{
A list containing the constraint matrix \code{Cmat}, and lower/upper bound vectors (\code{lb} and \code{ub}, respectively).
}
\description{
Methods to generate constraint matrices associated with \link[dlnm:crossbasis]{crossbasis} objects, allowing the fitting of constrained distributed lag linear and non-linear models (DLMs and DLNMs). Designed for use within the \link[=buildCmat]{constr} interface in \code{cirls}.
}
\details{
Constraints for DLMs and DLNMs apply to one of two dimensions of an exposure-lag-response association, specifically the exposure space (for \code{dim = "var"}) or the lag space (for \code{dim = "lag"}). Operationally, the functions generate a constraint matrix for the requested dimension of a \code{crossbasis} object, which is then expanded consistently in the other dimension. This allows the user to apply specific constraints to either the exposure-response association or the lag-response association, relying on the same \link[=shapeConstr]{shape} or \link[=boundConstr]{bound} methods used for other functions.
\subsection{Overall vs full surface}{

By default, when \code{dim = "var"}, the constraint is applied across the whole lag space. This means that the exposure-response will be constrained at every lag. When \code{overall} is switched to \code{TRUE} (allowed only when \code{dim = "var"}), the constraint will only hold for the overall cumulative association (namely the net effect summed across lags), while it can be violated for exposure-responses defined at specific lags. See \link[dlnm:crossreduce]{crossreduce} for more information.
}

\subsection{Slicing}{

The \code{slice} argument can be used to restrict the constraint on a specific range of the other dimension. For instance, when \code{dim = "var"}, setting \code{slice = c(0, 5)} (say) will enforce the constraint on the exposure-response relationship only over lags 0 to 5, meaning it could be violated at other lags (> 5 in this example). Note that the actual range for slicing depends on the specific basis functions used for the opposite dimension (see the \code{range} argument in \link{shapeConstr} for additional details). When \code{overall = TRUE}, slicing will only constraint the cumulative association over the range specific by \code{slice} (see above).
}

\subsection{Note}{

These methods for \link[dlnm:crossbasis]{crossbasis} objects only allow to specify a single constraint to one dimension of the exposure-lag-response association. However, the user can apply multiple constraints, possibly to both dimensions, by adding multiple terms for the same object in the \link[=buildCmat]{constr} formula (see Examples below).
}

\subsection{Warnings}{

In the current implementation, constraints applied to the lag dimension (when \code{dim = "lag"}) can be straightforwardly defined when the exposure-response function is linear, namely for simpler DLMs. For non-linear exposure-responses, the centering performed by \link[dlnm:crosspred]{crosspred} when predicting the exposure-lag-response relationship can break the constraint definition and return unpredictable results. Users are advised not to apply constraints in the lag dimension of complex DLNMs, or to do so with particular care.
}
}
\examples{
###############################################################################
# Constrained DLNM: var dimension

library(dlnm)
library(splines)

#----------------
# Association between temperature and mortality
#----------------

# Number of years and dow
ny <- length(unique(format(london$date, "\%Y")))

# Create a flexible crossbasis
cb <- crossbasis(london$tmean, lag = 21,
  argvar = list(fun = "bs", degree = 2, df = 10),
  arglag = list(fun = "ns", knots = logknots(21, df = 5)))

#----- Unconstrained model

# Fit the model
um <- glm(age0_64 ~ cb + ns(date, df = 7 * ny) + dow, data = london,
  family = "quasipoisson")
ucp <- crosspred(cb, um, cen = 20)

# Plot overall
plot(ucp, ptype = "overall", lwd = 2)

# Plot slices
plot(ucp, ptype = "slice", lag = 0, lwd = 2)
lines(ucp, ptype = "slice", lag = 5, col = 2, lwd = 2)
lines(ucp, ptype = "slice", lag = 15, col = 3, lwd = 2)
legend("topleft", legend = sprintf("lag = \%i", c(0, 5, 15)), col = 1:3,
  lty = 1)

#----- Constrained DLNM

# Fit the model with convexity constraint
scm <- glm(age0_64 ~ cb + ns(date, df = 7 * ny) + dow, data = london,
  family = "quasipoisson", method = "cirls.fit",
  constr = ~ shape(cb, shape = "cvx"))
scp <- crosspred(cb, scm, cen = 20)

# Plot slices
plot(scp, ptype = "slice", lag = 0, lwd = 2)
lines(scp, ptype = "slice", lag = 5, col = 2, lwd = 2)
lines(scp, ptype = "slice", lag = 15, col = 3, lwd = 2)
legend("topleft", legend = sprintf("lag = \%i", c(0, 5, 15)), col = 1:3,
  lty = 1)

#----- Overall only

# Fit the model with a constraint on overall only
ocm <- glm(age0_64 ~ cb + ns(date, df = 7 * ny) + dow, data = london,
  family = "quasipoisson", method = "cirls.fit",
  constr = ~ shape(cb, shape = "cvx", overall = TRUE))
ocp <- crosspred(cb, ocm, cen = 20)

# Plot slices: they don't necessarily respect the constraints
plot(ocp, ptype = "slice", lag = 0, lwd = 2)
lines(ocp, ptype = "slice", lag = 5, col = 2, lwd = 2)
lines(ocp, ptype = "slice", lag = 15, col = 3, lwd = 2)
legend("topleft", legend = sprintf("lag = \%i", c(0, 5, 15)), col = 1:3,
  lty = 1)

# Plot overall: constrained
plot(ocp, ptype = "overall", lwd = 2)
lines(ucp, ptype = "overall", col = 2, lwd = 2, ci = "lines")
legend("topleft", legend = c("Constrained", "Unconstrained"),
  col = 1:2, lty = 1:2)

#----- Slices

# Fit model
subcm <- glm(age0_64 ~ cb + ns(date, df = 7 * ny) + dow, data = london,
  family = "quasipoisson", method = "cirls.fit",
  constr = ~ shape(cb, shape = "cvx", slice = c(0, 1)))
subcp <- crosspred(cb, subcm, cen = 20)

# Plot slices: lags 0 and 5 are constrained but not lag 15
plot(subcp, ptype = "slice", lag = 0, lwd = 2)
lines(subcp, ptype = "slice", lag = 5, col = 2, lwd = 2)
lines(subcp, ptype = "slice", lag = 15, col = 3, lwd = 2)
legend("topleft", legend = sprintf("lag = \%i", c(0, 5, 15)), col = 1:3, lty = 1)


###############################################################################
# Constrained distributed lag models (DLM)

library(dlnm)
library(splines)

#----------------
# Association between PM2.5 and mortality
#----------------

# Number of years and dow
ny <- length(unique(format(london$date, "\%Y")))

# Create a flexible crossbasis
cb <- crossbasis(london$pm10, lag = 15, argvar = list(fun = "lin"),
  arglag = list(fun = "ns", knots = logknots(15, df = 4)))

#----- Unconstrained model

# Fit the model
um <- glm(age0_64 ~ cb + ns(date, df = 7 * ny) + dow, data = london,
  family = "quasipoisson")

# Plot lag-response function
ucp <- crosspred(cb, um, cen = 0, at = 10)
plot(ucp, ptype = "slices", lwd = 2, var = 10)

#----- Constrained to converge to zero

# Fit with bound constraint
# By default dim = "lag" (because var is linear) and value = 0
bm <- glm(age0_64 ~ cb + ns(date, df = 7 * ny) + dow, data = london,
  family = "quasipoisson", method = "cirls.fit", constr = ~ bound(cb))

# Plot
bcp <- crosspred(cb, bm, cen = 0, at = 10)
plot(bcp, ptype = "slices", lwd = 2, var = 10)

#----- Constrained to always be non-negative

# Fit with bound constraint
# By default dim = "lag" (because var is linear)
nnm <- glm(age0_64 ~ cb + ns(date, df = 7 * ny) + dow, data = london,
  family = "quasipoisson", method = "cirls.fit",
  constr = ~ shape(cb, shape = "pos"))

# Plot
nncp <- crosspred(cb, nnm, cen = 0, at = 10)
plot(nncp, ptype = "slices", lwd = 2, var = 10)

#----- Constrained to always decrease

# Fit with bound constraint
# By default dim = "lag" (because var is linear)
dm <- glm(age0_64 ~ cb + ns(date, df = 7 * ny) + dow, data = london,
  family = "quasipoisson", method = "cirls.fit",
  constr = ~ shape(cb, shape = "dec"))

# Plot
dcp <- crosspred(cb, dm, cen = 0, at = 10)
plot(dcp, ptype = "slices", lwd = 2, var = 10)

#----- Both decrease and towards zero

# Fit with bound constraint
# By default dim = "lag" (because var is linear)
dbm <- glm(age0_64 ~ cb + ns(date, df = 7 * ny) + dow, data = london,
  family = "quasipoisson", method = "cirls.fit",
  constr = ~ shape(cb, shape = "dec") + bound(cb))

# Plot
dbcp <- crosspred(cb, dbm, cen = 0, at = 10)
plot(dbcp, ptype = "slices", lwd = 2, var = 10)
}
\references{
Gasparrini, A., Armstrong, B., Kenward, M.G., 2010. Distributed lag non-linear models. \emph{Statistics in Medicine} \strong{29}, \strong{2224–2234}. \href{https://doi.org/10.1002/sim.3940}{DOI:10.1002/sim.3940}

Gasparrini, A., Armstrong, B., 2013. Reducing and meta-analysing estimates from distributed lag non-linear models. \emph{BMC Medical Research Methodology} \strong{13}, \strong{1}. \href{https://doi.org/10.1186/1471-2288-13-1}{DOI:10.1186/1471-2288-13-1}
}
\seealso{
\link[dlnm:crossbasis]{crossbasis} for the definition of DLMs and DLNMs as well as their dimensions. \link{shapeConstr} and \link{boundConstr} for generic constraint methods. \link{buildCmat} for how to specify constraints.
}
