% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warming.R
\docType{data}
\name{warming}
\alias{warming}
\title{Global temperature anomaly}
\format{
A data frame of 166 observations (years) and two? variables:
\describe{
\item{year}{The year.}
\item{decade}{The year rounded down to to the decade.}
\item{anomaly}{The difference between the global annual temperature and the average over 1961-1990.}
}
}
\source{
Jones, P.D., Parker, D.E., Osborn, T.J., Briffa, K.R., 2000. Global and Hemispheric Temperature Anomalies: Land and Marine Instrumental Records (1850 - 2015). \href{https://doi.org/10.3334/CDIAC/CLI.002}{DOI:10.3334/CDIAC/CLI.002}
}
\usage{
warming
}
\description{
Provides annual global temperature anomaly compared to the reference period 1961-1990 from 1850 to 2015
}
\examples{
###############################################################################
# Monotone strata levels with the warming dataset

### Fit the model

# Non-decreasing constraint on decadal strata
model <- glm(anomaly ~ decade, data = warming, method = "cirls.fit",
  cons = ~ shape(decade, "inc"))

# Plot result
plot(anomaly ~ year, data = warming, xlab = "", ylab = "Temperature anomaly")
lines(warming$year, predict(model), col = 2, lwd = 2)

### Extract results

# Coefficients and confidence intervals
betas <- coef(model)
v <- vcov(model)
cis <- confint(model)

# Degrees of freedom: represent number of strata change
# ?edf
edf(model)

### Compare with an unconstrained model

# Refit the model without constraints
umodel <- uncons(model)

# Add result
lines(warming$year, predict(umodel), col = 3, lwd = 2, lty = 2)

}
\keyword{datasets}
