% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundConstr.R, R/boundConstr.default.R,
%   R/boundConstr.factor.R
\name{boundConstr}
\alias{boundConstr}
\alias{bound}
\alias{boundConstr.default}
\alias{boundConstr.factor}
\title{Boundary constraints}
\usage{
boundConstr(x, ...)

\method{boundConstr}{default}(x, value = 0, side = c("right", "left",
  "both"), sm = 1, intercept = FALSE, ...)

\method{boundConstr}{factor}(x, value = 0, side = "right", thr = NULL,
  sm = 1, intercept = FALSE, ...)
}
\arguments{
\item{x}{An object representing a design matrix of predictor variables, typically basis functions. See Details for supported objects.}

\item{...}{Additional parameters passed to or from other methods. Includes \code{value} and \code{side} for most methods.}

\item{value}{A numerical indicating the boundary constraint value. Default to 0.}

\item{side}{Character indicating the side on which the constraint applies. One of \code{"right"} (the default), \code{"left"} or \code{"both"}.}

\item{sm}{A positive integer indicating the degree of smoothness for the constraint. In the default method indicates the number of coefficients to constrain to \code{value}, but can have a slightly different interpretation in other methods.}

\item{intercept}{For the default and factor method, a logical value indicating if the design matrix includes an intercept. In most cases, it will be automatically extracted from \code{x}, but this argument can be used to override it.}

\item{thr}{For the factor method, an integer indicating a level from which the association should be equal to \code{value}. Cannot be used with \code{side = both}. This parameter has a slightly different interpretation for other methods.}
}
\value{
A list containing the constraint matrix \code{Cmat}, and lower/upper bound vectors (\code{lb} and \code{ub}, respectively).
}
\description{
Generic function to build a constraint matrix and bound vectors for constraining the boundary of \link[=boundConstr.bs]{B-splines} and other nonlinear functions to a given value. It is designed to be used in the \link[=buildCmat]{constr} formula interface. It allows methods for a wide range of regression terms.
}
\details{
Enforcing boundary constraints amounts to imposing an equality constraint on the \code{sm} first (if \code{side = left}), last (if \code{side = right}), or both (if \code{side = both}) coefficients of the basis matrix \code{x}. The equality constraint sets the bounds to \code{lb = ub = value}.
\subsection{Usage}{

The recommended usage is to use this function through a call to \code{bound} on a term in the \link[=buildCmat]{constr} formula interface. This method is then called internally to create the constraint matrix and bound vectors. However, \code{boundConstr} can also be called directly on a matrix-like object to manually build or inspect the constraint matrix.

All methods internally rely on the default method for general matrices. Unless specified, all methods use the same parameters as the default one, which are passed through \code{...}. The only exception is \code{intercept} which is often inferred from the \code{x} object itself. In a typical usage in which \code{boundConstr} is called from the \code{constr} argument, \code{intercept} is automatically determined from the \link[stats:glm]{glm} formula.
}

\subsection{The \code{sm} parameter}{

The \code{sm} parameter indicates the number of coefficients on the left/right that are constrained to be equal to \code{value} and can be interpreted as a smoothness degree for the boundary constraint. The default is to constraint only one coefficient. Note that \code{sm} has a slightly different interpretation when \code{x} represents \link[=boundConstr.bs]{spline bases}.
}

\subsection{Available methods}{

In addition to the default method, \code{boundConstr} currently supports methods for several classes. The full list can also be consulted through \code{methods(boundConstr)}.
\subsection{Categorical variables}{
\itemize{
\item \link{factor}: for \code{factor} objects. Extract the \link[stats:contrasts]{contrasts} to define the constraint matrix. Here the \code{intercept} argument has the same interpretation as in the default method, i.e. if set to \code{TRUE} it means the \code{glm} model does not include an intercept externally to the factor. Note that, in this case, a simple dummy coding is done in R.
\item \link[=boundConstr.strata]{strata}: Indicator variables defining strata from the \link[dlnm:dlnm]{dlnm} package.
}
}

\subsection{Splines}{
\itemize{
\item \link[=boundConstr.bs]{bs, ns, ps}: Specify boundary constraints on various flavors of B-splines.
}
}

\subsection{Distributed-lag linear and nonlinear models (DLNM)}{
\itemize{
\item \link[=boundConstr.onebasis]{onebasis}: General method for basis functions generated in the package. Internally calls the relevant method for the specified basis.
\item \link[=boundConstr.crossbasis]{crossbasis}: Most exhaustive method to constrain DLNMs. See its dedicated \link[=boundConstr.crossbasis]{help page}.
}
}

}
}
\examples{
###############################################################################
# Bound constraints example: warming dataset

library(splines)

# Force the fit to reach the last value
# Remove the intercept, so that it is included in the constraint
model <- glm(anomaly ~ 0 + decade, data = warming, method = "cirls.fit",
  constr = ~ bound(decade, value = 0.75))

# Same but with splines
basis <- ns(warming$year, df = 10)
splmodel <- glm(anomaly ~ 0 + basis, data = warming, method = "cirls.fit",
  constr = ~ bound(basis, value = 0.75))

# Plot result
plot(anomaly ~ year, data = warming, xlab = "", ylab = "Temperature anomaly")
lines(warming$year, predict(model), col = 2, lwd = 2)
lines(warming$year, predict(splmodel), col = 3, lwd = 2)
}
\seealso{
\link{buildCmat} detailing the \code{constr} interface.
}
