% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCmat.R
\name{checkCmat}
\alias{checkCmat}
\title{Check and reduce constraint matrix}
\usage{
checkCmat(Cmat, lb = NULL, ub = NULL, reduce = TRUE, warn = TRUE)
}
\arguments{
\item{Cmat}{A constraint matrix.}

\item{lb, ub}{Bound vectors. If provided, should be consistent with \code{Cmat}.}

\item{reduce}{If TRUE, returns reduced \code{Cmat}, \code{lb} and \code{ub} with redundant constraints removed. Otherwise, they are returned as provided.}

\item{warn}{Whether to warn the user when redundant and equality constraints are found.}
}
\value{
A list with the following elements:
\item{redundant}{Logical vector indicating redundant constraints expressed in the rows \code{Cmat}.}
\item{equality}{An integer vector indicating the groups of underlying equality constraints in the provided \code{Cmat}. Zero values indicate that the row is not part of any underlying equality constraint and higher values indicate which equality constraint it is part of.}
\item{Cmat/lb/ub}{The reduced constraints when \code{reduce = TRUE} or the provided constraints otherwise.}
}
\description{
Checks whether some constraints are redundant and can be removed, and whether some constraints can be reduced into an equality constraint. If requested, redundant constraints are removed, but not equality ones.
}
\details{
The user typically does not need to use \code{checkCmat} as it is internally called by \link[=buildCmat]{buildCmat} and in some Constr functions used to create \code{Cmat}/\code{lb}/\code{ub} . However, it might be useful in the case of inconsistent constraints or underlying equality constraints when users are building such objects directly.
\subsection{Irreducibility}{

\code{Cmat} is irreducible if there is no \emph{redundant} constraint, and no \emph{underlying equality} constraint.

A row in \code{Cmat} is \emph{redundant} if it can be expressed as a \emph{positive} linear combination of other rows. When it happens, it means the corresponding constraint is actually implicitly included in other constraints, and can be dropped without affecting the problem. For instance, rows that only contain zeros are considered redundant.

When it exists a linear combination of rows in \code{Cmat} that results in a null vector, it means that there is an \emph{underlying equality} constraint. In that case, it means the corresponding rows can be reduced to a single equality constraint. Note that underlying constraints are left as they are in the returned \code{Cmat} even when \code{reduce = TRUE}, and it is left to the user to figure out whether to reduce them.
}
}
\note{
\code{checkCmat} works when only \code{Cmat} is provided. \code{lb}/\code{ub} can be provided to conveniently be passed and reduced in the case of redundant constraints, but they must be consistent with \code{Cmat}.
}
\examples{
###################################################
# Example of reducible matrix

# Constraints: successive coefficients should increase and be convex
# Intuitively, if the first two coefficients increase,
# then convexity forces the rest to increase which means there is redundancy
p <- 5
cmatic <- rbind(
  shapeConstr(matrix(NA, 0, p), shape = "inc")$Cmat, # Increasing
  shapeConstr(matrix(NA, 0, p), shape = "cvx")$Cmat # Convex
)

# Checking indicates that some constraints are redundant
# Returns reduced matrix and a warning
checkCmat(cmatic)

# Compare without removing the redundant constraints
checkCmat(cmatic, reduce = FALSE)

# Note that this is silently done when both "inc" and "cvx" are provided
shapeConstr(matrix(NA, 0, p), shape = c("inc", "cvx"))$Cmat

###################################################
# Example of irreducible matrix

# Constraints: coefficients form an S-shape
p <- 4
cmats <- rbind(
  diag(p)[1,], # positive
  diff(diag(p))[c(1, p - 1),], # Increasing at both end
  diff(diag(p), diff = 2)[1:(p/2 - 1),], # First half convex
  -diff(diag(p), diff = 2)[(p/2):(p-2),] # second half concave
)

# Note, this matrix is not of full row rank
qr(t(cmats))$rank
all.equal(cmats[2,] + cmats[4,] - cmats[5,], cmats[3,])

# However, it is irreducible: all constraints are necessary
checkCmat(cmats)

###################################################
# Example of underlying equality constraint

# Contraint: Parameters sum is >= 0 and sum is <= 0
cmateq <- rbind(rep(1, 3), rep(-1, 3))

# Checking indicates that both constraints imply equality constraint
checkCmat(cmateq)
}
\references{
Meyer, M.C., 1999. An extension of the mixed primal–dual bases algorithm to the case of more constraints than dimensions. \emph{Journal of Statistical Planning and Inference} \strong{81}, 13–31. \href{https://doi.org/10.1016/S0378-3758(99)00025-7}{DOI:10.1016/S0378-3758(99)00025-7}
}
\seealso{
\link[=buildCmat]{buildCmat}
}
