% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cirls-package.R
\docType{package}
\name{cirls-package}
\alias{cirls}
\alias{cirls-package}
\title{cirls: Constrained Iteratively Reweighted Least Squares}
\description{
The package \code{cirls} provides routines to fit \strong{generalised linear models (GLM) with linear constraints on the coefficients}. These routines implement a general framework to perform a range of constraint-based methods, including common applications such as shape-constrained regression or compositional regression. However, the package allows users to specify \emph{any} custom linear constraint for more niche applications. Functions in the \code{cirls} package include fitting procedures as well as dedicated methods to perform inference and model selection.

The statistical framework is implemented by embedding constrained optimisation procedures in \link[stats:glm]{glm}, so that users can flexibly perform constraint-based analyses with simple calls to this standard R function. The user can use the standard \code{glm} ecosystem on constrained model, augmented with dedicated methods for \code{cirls} objects.
}
\section{Methodological framework}{

Applying linear constraints to GLMs entails restricting the fitted coefficients \eqn{\beta} to respect specific linear combinations. These constraints can represent assumptions on the coefficients or enforce identifiability. More formally, the fitted GLM is subject to:
\deqn{ \mathbf{l} \leq \mathbf{C\beta} \leq \mathbf{u} }
where \eqn{\mathbf{C}} is a matrix encoding the constraints, with each row representing a single constraint and each column representing a regressor in the design matrix of the model. \eqn{\mathbf{l}} and \eqn{\mathbf{u}} are lower and upper bounds for the constraints, respectively.

The model is fitted by \link[=cirls.fit]{constrained iteratively-reweighted least-squares} (CIRLS), which extends the classical IRLS algorithm to ensure the constraints specified by \eqn{\mathbf{C}}, \eqn{\mathbf{l}}, and \eqn{\mathbf{u}} are respected.

Linearly constrained coefficients follow a \strong{truncated multivariate Normal} (TMVN) distribution, \emph{i.e.} a multivariate normal that respects the constraints specified by \eqn{\mathbf{C}}, \eqn{\mathbf{l}} and \eqn{\mathbf{u}}. Inference is conducted by \link[=simulCoef]{simulating} from the TMVN distribution of fitted coefficients, and the output can then be used to compute useful summaries like the \link[=vcov.cirls]{variance-covariance matrix} and \link[=confint.cirls]{confidence intervals}.
}

\section{Usage}{

\subsection{Fitting the model}{

The \code{cirls} package has been developed as a \strong{plug-in} to \link[stats:glm]{glm}. The main function is \link{cirls.fit}, designed to be used within \link[stats:glm]{glm} through the argument \code{method = "cirls.fit"}. This will fit the GLM using specific CIRLS algorithms instead of the default IRLS algorithm. All arguments of \link[stats:glm]{glm} such as \code{formula}, \code{data} and \code{family} retain the same usage. Additional arguments specific to \code{cirls} are passed through the ellipsis \code{...} or the \code{control} argument (\emph{but not both at the same time}). If any parameter is passed through \code{control}, then \code{...} will be ignored. See \link{cirls.control} for the full list of additional arguments and examples below for how to use additional functions in the package.
}

\subsection{Specifying constraints}{

Constraints can be specified through arguments \code{constr} and/or \code{Cmat}/\code{lb}/\code{ub}. The \link[=buildCmat]{constr} argument provides a user-friendly way to pass common built-in constraints through a formula and is the recommended way for new users. See the list of available built-in constraints below. Alternatively, the arguments \link[=buildCmat]{Cmat/lb/ub} allow more flexibility in passing constraints. These arguments take either a full constraint matrix (\code{Cmat}) and bound vectors (\code{lb} and \code{ub}), or named lists to specify a constrained matrix and bound vectors for specific terms in the model. See the help for \link{buildCmat} for more specific details.
}
}

\section{Functions included in the package}{

\link{cirls.fit} is the workhorse performing the fitting, and is controlled by \link{cirls.control}. The function returns an object of class \code{cirls} which inherits from class \code{glm}. This means that all methods available for usual \code{glm} objects can also be used on a \code{cirls} object. This typically includes \link[stats:coef]{coef} and \link[stats:summary.glm]{summary}, for instance. However, some methods are supplanted by new methods that are specific to \code{cirls} objects. The sections below provides a comprehensive illustration of such specific methods, including an index of the functions.
\subsection{Inference}{

Inference for \code{cirls} objects includes computing the \link[=vcov.cirls]{variance-covariance matrix} and \link[=confint.cirls]{confidence intervals} for constrained (and unconstrained) coefficients. These functions build on \link{simulCoef}, which generates coefficient vectors based on the TMVN distribution of fitted coefficients, and can be used to compute other summaries.
}

\subsection{Model selection}{

Degrees of freedom of a fitted \code{cirls} object can be computed through the \link{edf} function. A specific \link[=logLik.cirls]{logLik} method extracts the log-likelihood of the model with the right degrees of freedom in its \code{df} attribute. This can be used by \link[stats:AIC]{AIC} and \link[stats:BIC]{BIC} with appropriate degrees of freedom.
}

\subsection{Other functions}{

\link{uncons} refits the model removing some or all of the constraints, acting similarly to the \link[stats:update]{update} function although focusing on the constraints.

\link{buildCmat} and \link{reduceCons} are convenience functions that are mainly used internally in \link{cirls.fit}. They are not meant to be called directly, but can be useful for advanced users to build and/or check constraint matrices. Specifically, \link{buildCmat} builds a full constraint matrix from the \code{constr} and/or \code{Cmat}/\code{lb}/\code{ub} arguments and a model frame. \link{reduceCons} checks whether a constraint matrix can be reduced by detecting redundant and underlying equality constraints. The help pages of both functions also provide technical details on constraints and constraint matrices.
}
}

\section{Index}{

\subsection{Built-in constraints}{

These constraints can be used in the \code{constr} formula as functions applied to specific terms:
\itemize{
\item \link[=shapeConstr]{shape}: constraining the shape of an association. Typically used with spline bases or factors.
\item \link[=zerosumConstr]{zerosum}: specify that a group of coefficients should sum to zero. Typically used for compositional regression.
\item \link[=boundConstr]{bound}: boundary constraint for an association. Typically used with spline bases or factors.
}
}

\subsection{Methods specific to \code{cirls} objects}{

The following supplant \code{glm} methods:
\itemize{
\item \link[=vcov.cirls]{vcov}: compute the variance-covariance matrix of coefficients.
\item \link[=confint.cirls]{confint}: compute confidence intervals for constrained coefficients.
\item \link[=logLik.cirls]{logLik}: extract the log-likelihood of a fitted object, assigning the correct \link[=edf]{degrees of freedom}.
}
}
}

\section{Datasets}{

\itemize{
\item \link{fgl}: Measurement of forensic glass fragments
\item \link{london}: Daily mortality, temperature, and pollution data in London
\item \link{warming}: Global temperature anomaly
}
}

\examples{
###############################################################################
# Non-negative coefficient example with the london dataset

library(splines)

### Association between Ozone and mortality
# Nonnegative constraint on Ozone
model <- glm(death ~ o3, data = london, family = "quasipoisson",
  method = "cirls.fit", cons = ~ shape(o3, "pos"))

# Coefficient and confidence interval
coef(model)[2]
confint(model)[2,]

# Comparing to an unconstrained model
umodel <- uncons(model)
coef(umodel)[2]
confint(umodel)[2,]
###############################################################################
# Monotone strata levels with the warming dataset

### Fit the model

# Non-decreasing constraint on decadal strata
model <- glm(anomaly ~ decade, data = warming, method = "cirls.fit",
  cons = ~ shape(decade, "inc"))

# Plot result
plot(anomaly ~ year, data = warming, xlab = "", ylab = "Temperature anomaly")
lines(warming$year, predict(model), col = 2, lwd = 2)

### Extract results

# Coefficients and confidence intervals
betas <- coef(model)
v <- vcov(model)
cis <- confint(model)

# Degrees of freedom: represent number of strata change
# ?edf
edf(model)

### Compare with an unconstrained model

# Refit the model without constraints
umodel <- uncons(model)

# Add result
lines(warming$year, predict(umodel), col = 3, lwd = 2, lty = 2)

##############################################################################
# Monotone splines with the warming dataset

library(splines)

# Define a natural spline basis
basis <- ns(warming$year, df = 10)

### Constrained model

# Non-decreasing splines
model <- glm(anomaly ~ basis, data = warming, method = "cirls.fit",
  cons = ~ shape(basis, "inc"))

# Plot result
plot(anomaly ~ year, data = warming, xlab = "", ylab = "Temperature anomaly")
lines(warming$year, predict(model), col = 2, lwd = 2)

### Unconstrained model
umodel <- uncons(model)
lines(warming$year, predict(umodel), col = 3, lwd = 2, lty = 2)

###############################################################################
# Compositional regression example with the fgl dataset

# Select the composition and use log
x <- as.matrix(fgl[,2:9])
z <- log(x)

# Fit model
model <- glm(RI ~ z, data = fgl, method = "cirls.fit",
  cons = ~ zerosum(z))

# Coefficients sum to 1
coef(model)
sum(coef(model)[-1])

# Confidence intervals
confint(model)

}
\references{
Masselot, P., Nenon, D., Vanoli, J., Chalabi, Z., Gasparrini, A., 2025. Estimation and inference in generalised linear models with constrained iteratively-reweighted least squares. \emph{ArXiv preprint}. \href{https://doi.org/10.48550/arXiv.2509.18406}{DOI:10.48550/arXiv.2509.18406}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PierreMasselot/cirls}
}

}
\author{
\strong{Maintainer}: Pierre Masselot \email{pierre.masselot@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-7326-1290}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Pierre Masselot \email{pierre.masselot@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-7326-1290}{ORCID}) [copyright holder]
  \item Antonio Gasparrini \email{antonio.gasparrini@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-2271-3568}{ORCID})
}

}
