% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cirls.control.R
\name{cirls.control}
\alias{cirls.control}
\title{Parameters controlling CIRLS fitting}
\usage{
cirls.control(epsilon = 1e-08, maxit = 25, trace = FALSE, Cmat = NULL,
  lb = 0L, ub = Inf, qp_solver = "osqp", qp_pars = list())
}
\arguments{
\item{epsilon}{Positive convergence tolerance \eqn{\epsilon}; the algorithm converges when \eqn{|dev - dev_{old}|/(|dev| + 0.1) < \epsilon}.}

\item{maxit}{Integer giving the maximal number of CIRLS iterations.}

\item{trace}{Logical indicating if output should be produced for each iteration.}

\item{Cmat}{Constraint matrix specifying the linear constraints applied to coefficients. Can also be provided as a list of matrices for specific terms.}

\item{lb, ub}{Lower and upper bound vectors for the linear constraints. Identical values in \code{lb} and \code{ub} identify equality constraints. Recycled if length is different than the number of constraints defined by \code{Cmat}.}

\item{qp_solver}{The quadratic programming solver. One of \code{"osqp"}, \code{"quadprog"} or \code{"coneproj"}.}

\item{qp_pars}{List of parameters specific to the quadratic programming solver. See respective packages help.}
}
\value{
A named list containing arguments to be used in \link[=cirls.fit]{cirls.fit}.
}
\description{
Internal function controlling the \link[stats:glm]{glm} fit with linear constraints. Typically only used internally by \link[=cirls.fit]{cirls.fit}, but may be used to construct a control argument.
}
\details{
The \code{control} argument of \link[stats:glm]{glm} is by default passed to the \code{control} argument of \link[=cirls.fit]{cirls.fit}, which uses its elements as arguments for \link[=cirls.control]{cirls.control}: the latter provides defaults and sanity checking. The control parameters can alternatively be passed through the \code{...} argument of \link[stats:glm]{glm}. See \link[stats:glm.control]{glm.control} for details on general GLM fitting control, and \link[=cirls.fit]{cirls.fit} for details on arguments specific to constrained GLMs.
}
\seealso{
the main function \link[=cirls.fit]{cirls.fit}, and \link[stats:glm.control]{glm.control}.
}
