% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgl.R
\docType{data}
\name{fgl}
\alias{fgl}
\title{Measurement of forensic glass fragments}
\format{
A data frame of 214 observations and 10 variables. Apart from \code{RI} and \code{type} all other variables represent a percentage of the given chemical component.
\describe{
\item{RI}{Refractive index; more precisely the refractive index is 1.518xxxx.}
\item{Na}{Sodium}
\item{Mg}{Manganese}
\item{Al}{Aluminium}
\item{Si}{Silicon}
\item{K}{Potassium}
\item{Ca}{Calcium}
\item{Ba}{Barium}
\item{Fe}{Iron}
\item{type}{The fragments were originally classed into seven types, one of which was absent in this dataset. The categories which occur are window float glass (\code{WinF}: 70), window non-float glass (\code{WinNF}: 76), vehicle window glass (\code{Veh}: 17), containers (\code{Con}: 13), tableware (\code{Tabl}: 9) and vehicle headlamps (\code{Head}: 29)}
}
}
\source{
Dataset exctracted from the \code{MASS} package.

Venables, W. N. and Ripley, B. D. (2002) \emph{Modern Applied Statistics with S}. Fourth edition. Springer.
}
\usage{
fgl
}
\description{
Chemical composition of fragments of glass collected in forensic work.
}
\examples{
###############################################################################
# Compositional regression example with the fgl dataset

# Add a small increment to zeros
x <- as.matrix(fgl[,2:9])
x[x == 0] <- min(x[x > 0]) / 2

# Normalise to unit sum and take log
z <- log(x / rowSums(x))

# Fit model
model <- glm(RI ~ z, data = fgl, method = "cirls.fit",
  cons = ~ zerosum(z))

# Coefficients and confidence intervals
coef(model)
confint(model)
}
\keyword{datasets}
