% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundConstr.crossbasis.R,
%   R/shapeConstr.crossbasis.R
\name{boundConstr.crossbasis}
\alias{boundConstr.crossbasis}
\alias{shapeConstr.crossbasis}
\title{Constraints for distributed lag linear and nonlinear models}
\usage{
\method{boundConstr}{crossbasis}(x, dim = NULL, slice = NULL,
  overall = FALSE, ...)

\method{shapeConstr}{crossbasis}(x, dim = NULL, slice = NULL,
  overall = FALSE, ...)
}
\arguments{
\item{x}{A \code{crossbasis} object.}

\item{dim}{One of \code{"var"} or \code{"lag"}, the dimension on which the constraint will be applied. If \code{NULL}, the default is to apply the constraint on the \code{"var"} dimension if it is \emph{nonlinear} (i.e. has more than one degree of freedom), or to apply it to the \code{"lag"} dimension otherwise (i.e. when a \emph{linear} distributed lag model is specified).}

\item{slice}{A numeric vector of length 2 restricting the constraint on a specific range of the \emph{other} dimension. By default no slicing is performed. See details.}

\item{overall}{Only when \code{dim = "var"}, logical indicating whether the constraint should be applied only on the overall cumulative association or across all lags. Note that \code{slice} can be used to define the constraint over a restricted cumulative lag range.}

\item{...}{Parameters specific to the type of constraint to be applied. Includes for instance \code{shape} for \link[=shapeConstr]{shapeConstr}, or \code{value} for \link[=boundConstr]{boundConstr}. See the main help page of the relevant generic.}
}
\value{
A list containing the constraint matrix \code{Cmat}, and lower/upper bound vectors (\code{lb} and \code{ub}, respectively).
}
\description{
Methods to generate constraint matrices associated with \link[dlnm:crossbasis]{crossbasis} objects, allowing the fitting of constrained distributed-lag nonlinear models (DLNM). Designed for use within the \link[=buildCmat]{constr} interface in \code{cirls}. \strong{At the moment, constraining can only be performed on a single dimension at a time.}
}
\details{
Constraint matrices for \code{crossbasis} objects are built by generating the marginal constraint matrix for the requested dimension (through the \code{dim} argument) and expanding it through a Kronecker product with a (potentially modified) identity matrix of the dimension of the opposite dimension. This identity matrix can be modified to account for specificities of the basis for the other dimension (e.g. necessary for \code{ns}).
\subsection{Overall vs full surface}{

By default, the constraint is applied over the whole surface. This means that if, say, the \code{var} dimension is shape-constrained, then the shape will hold true for any lag. When \code{overall} is switched to \code{TRUE} (only when \code{dim = "var"}), the constrain will hold for the \emph{overall cumulative association} only (see \link[dlnm:crossreduce]{crossreduce}). This means the constraint could be violated at specific lags.
}

\subsection{Slicing}{

The \code{slice} argument can be used to restrict the constraint on a specific range of the other dimension. For instance, when \code{dim = "var"}, setting \code{slice = c(0, 5)} (say) will enforce the constraint only over lags 0 to 5, meaning it could be violated at other lags (> 5 in this example). Note that the specific range for slicing depends on the specific basis for the opposite dimension. In the previous example, if the lag dimension is specified by splines, the constraint will be enforced between the closest knots that contain the range specified by \code{slice}. See the \code{range} argument in \link[=shapeConstr]{shapeConstr} for additional details.

When \code{overall = TRUE}, slicing will only constraint the cumulative association \emph{over the range specific by \code{slice}}.
}

\subsection{Note}{

These methods only allow to specify a single constraint on \strong{one dimension at a time}. Specifying several constraints can be done by adding terms in the \link[=buildCmat]{constr} formula.

At the moment, \strong{when the var dimension is nonlinear, it is not recommended to constrain the lag dimension}. This is because the centering performed by \code{crosspred} can impact the constraint definition.
}
}
\examples{
###############################################################################
# Constrained DLNM: var dimension

library(dlnm)
library(splines)

#----------------
# Association between temperature and mortality
#----------------

# Number of years and dow
ny <- length(unique(format(london$date, "\%Y")))

# Create a flexible crossbasis
cb <- crossbasis(london$tmean, lag = 21,
  argvar = list(fun = "bs", degree = 2, df = 10),
  arglag = list(fun = "ns", knots = logknots(21, df = 5)))

#----- Unconstrained model

# Fit the model
um <- glm(age0_64 ~ cb + ns(date, df = 7 * ny) + dow, data = london,
  family = "quasipoisson")
ucp <- crosspred(cb, um, cen = 20)

# Plot overall
plot(ucp, ptype = "overall", lwd = 2)

# Plot slices
plot(ucp, ptype = "slice", lag = 0, lwd = 2)
lines(ucp, ptype = "slice", lag = 5, col = 2, lwd = 2)
lines(ucp, ptype = "slice", lag = 15, col = 3, lwd = 2)
legend("topleft", legend = sprintf("lag = \%i", c(0, 5, 15)), col = 1:3,
  lty = 1)

#----- Constrained DLNM

# Fit the model with convexity constraint
scm <- glm(age0_64 ~ cb + ns(date, df = 7 * ny) + dow, data = london,
  family = "quasipoisson", method = "cirls.fit",
  constr = ~ shape(cb, shape = "cvx"))
scp <- crosspred(cb, scm, cen = 20)

# Plot slices
plot(scp, ptype = "slice", lag = 0, lwd = 2)
lines(scp, ptype = "slice", lag = 5, col = 2, lwd = 2)
lines(scp, ptype = "slice", lag = 15, col = 3, lwd = 2)
legend("topleft", legend = sprintf("lag = \%i", c(0, 5, 15)), col = 1:3,
  lty = 1)

#----- Overall only

# Fit the model with a constraint on overall only
ocm <- glm(age0_64 ~ cb + ns(date, df = 7 * ny) + dow, data = london,
  family = "quasipoisson", method = "cirls.fit",
  constr = ~ shape(cb, shape = "cvx", overall = TRUE))
ocp <- crosspred(cb, ocm, cen = 20)

# Plot slices: they don't necessarily respect the constraints
plot(ocp, ptype = "slice", lag = 0, lwd = 2)
lines(ocp, ptype = "slice", lag = 5, col = 2, lwd = 2)
lines(ocp, ptype = "slice", lag = 15, col = 3, lwd = 2)
legend("topleft", legend = sprintf("lag = \%i", c(0, 5, 15)), col = 1:3,
  lty = 1)

# Plot overall: constrained
plot(ocp, ptype = "overall", lwd = 2)
lines(ucp, ptype = "overall", col = 2, lwd = 2, ci = "lines")
legend("topleft", legend = c("Constrained", "Unconstrained"),
  col = 1:2, lty = 1:2)

#----- Slices

# Fit model
subcm <- glm(age0_64 ~ cb + ns(date, df = 7 * ny) + dow, data = london,
  family = "quasipoisson", method = "cirls.fit",
  constr = ~ shape(cb, shape = "cvx", slice = c(0, 1)))
subcp <- crosspred(cb, subcm, cen = 20)

# Plot slices: lags 0 and 5 are constrained but not lag 15
plot(subcp, ptype = "slice", lag = 0, lwd = 2)
lines(subcp, ptype = "slice", lag = 5, col = 2, lwd = 2)
lines(subcp, ptype = "slice", lag = 15, col = 3, lwd = 2)
legend("topleft", legend = sprintf("lag = \%i", c(0, 5, 15)), col = 1:3, lty = 1)


###############################################################################
# Constrained distributed lag models (DLM)

library(dlnm)
library(splines)

#----------------
# Association between PM2.5 and mortality
#----------------

# Number of years and dow
ny <- length(unique(format(london$date, "\%Y")))

# Create a flexible crossbasis
cb <- crossbasis(london$pm10, lag = 15, argvar = list(fun = "lin"),
  arglag = list(fun = "ns", knots = logknots(15, df = 4)))

#----- Unconstrained model

# Fit the model
um <- glm(age0_64 ~ cb + ns(date, df = 7 * ny) + dow, data = london,
  family = "quasipoisson")

# Plot lag-response function
ucp <- crosspred(cb, um, cen = 0, at = 10)
plot(ucp, ptype = "slices", lwd = 2, var = 10)

#----- Constrained to converge to zero

# Fit with bound constraint
# By default dim = "lag" (because var is linear) and value = 0
bm <- glm(age0_64 ~ cb + ns(date, df = 7 * ny) + dow, data = london,
  family = "quasipoisson", method = "cirls.fit", constr = ~ bound(cb))

# Plot
bcp <- crosspred(cb, bm, cen = 0, at = 10)
plot(bcp, ptype = "slices", lwd = 2, var = 10)

#----- Constrained to always be non-negative

# Fit with bound constraint
# By default dim = "lag" (because var is linear)
nnm <- glm(age0_64 ~ cb + ns(date, df = 7 * ny) + dow, data = london,
  family = "quasipoisson", method = "cirls.fit",
  constr = ~ shape(cb, shape = "pos"))

# Plot
nncp <- crosspred(cb, nnm, cen = 0, at = 10)
plot(nncp, ptype = "slices", lwd = 2, var = 10)

#----- Constrained to always decrease

# Fit with bound constraint
# By default dim = "lag" (because var is linear)
dm <- glm(age0_64 ~ cb + ns(date, df = 7 * ny) + dow, data = london,
  family = "quasipoisson", method = "cirls.fit",
  constr = ~ shape(cb, shape = "dec"))

# Plot
dcp <- crosspred(cb, dm, cen = 0, at = 10)
plot(dcp, ptype = "slices", lwd = 2, var = 10)

#----- Both decrease and towards zero

# Fit with bound constraint
# By default dim = "lag" (because var is linear)
dbm <- glm(age0_64 ~ cb + ns(date, df = 7 * ny) + dow, data = london,
  family = "quasipoisson", method = "cirls.fit",
  constr = ~ shape(cb, shape = "dec") + bound(cb))

# Plot
dbcp <- crosspred(cb, dbm, cen = 0, at = 10)
plot(dbcp, ptype = "slices", lwd = 2, var = 10)
}
\references{
Gasparrini, A., Armstrong, B., 2013. Reducing and meta-analysing estimates from distributed lag non-linear models. \emph{BMC Medical Research Methodology} \strong{13}, \strong{1}. \href{https://doi.org/10.1186/1471-2288-13-1}{DOI:10.1186/1471-2288-13-1}

Gasparrini, A., Armstrong, B., Kenward, M.G., 2010. Distributed lag non-linear models. \emph{Statistics in Medicine} \strong{29}, \strong{2224–2234}. \href{https://doi.org/10.1002/sim.3940}{DOI:10.1002/sim.3940}
}
\seealso{
\link[=shapeConstr]{shapeConstr}, \link[=boundConstr]{boundConstr} for the generic method. \link[=buildCmat]{buildCmat} for how to specify constraints.
}
