% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundConstr.crossbasis.R,
%   R/shapeConstr.crossbasis.R
\name{boundConstr.crossbasis}
\alias{boundConstr.crossbasis}
\alias{shapeConstr.crossbasis}
\title{Constraints for distributed lag linear and non-linear models}
\usage{
\method{boundConstr}{crossbasis}(x, dim = "var", overall = FALSE,
  slice = NULL, ...)

\method{shapeConstr}{crossbasis}(x, dim = "var", overall = FALSE,
  slice = NULL, ...)
}
\arguments{
\item{x}{A \code{crossbasis} object.}

\item{dim}{Either \code{"var"} (the default) or \code{"lag"}. This is the dimension on which the constraint will be applied, constraining either the exposure-response or lag-response relationship, respectively.}

\item{overall}{Only when \code{dim = "var"}, logical indicating whether the constraint should be applied only on the overall cumulative association or for each specific lag.}

\item{slice}{A numeric vector of length 2 restricting the constraint on a specific range of the \emph{other} dimension, namely the lag (for \code{dim = "var"}) or exposure (\code{dim = "lag"}) space. By default, no slicing is performed. See Details.}

\item{...}{Parameters specific to the type of constraint to be applied. Includes for instance \code{shape} for \link{shapeConstr}, or \code{value} for \link{boundConstr}. See the main help page of the relevant generic method.}
}
\value{
A list containing the constraint matrix \code{Cmat}, and lower/upper bound vectors (\code{lb} and \code{ub}, respectively).
}
\description{
Methods to generate constraint matrices associated with \link[dlnm:crossbasis]{crossbasis} objects, allowing the fitting of constrained distributed lag linear and non-linear models (DLMs and DLNMs). Designed for use within the \link[=buildCmat]{constr} interface in \code{cirls}.
}
\details{
Constraints for DLMs and DLNMs apply to one of two dimensions of an exposure-lag-response association, specifically the exposure space (for \code{dim = "var"}) or the lag space (for \code{dim = "lag"}). Operationally, the functions generate a constraint matrix for the requested dimension of a \code{crossbasis} object, which is then expanded consistently in the other dimension. This allows the user to apply specific constraints to either the exposure-response association or the lag-response association, relying on the same \link[=shapeConstr]{shape} or \link[=boundConstr]{bound} methods used for other functions.
\subsection{Overall vs full surface}{

By default, when \code{dim = "var"}, the constraint is applied across the whole lag space. This means that the exposure-response will be constrained at every lag. When \code{overall} is switched to \code{TRUE} (allowed only when \code{dim = "var"}), the constraint will only hold for the overall cumulative association (namely the net effect summed across lags), while it can be violated for exposure-responses defined at specific lags. See \link[dlnm:crossreduce]{crossreduce} for more information.
}

\subsection{Slicing}{

The \code{slice} argument can be used to restrict the constraint on a specific range of the other dimension. For instance, when \code{dim = "var"}, setting \code{slice = c(0, 5)} (say) will enforce the constraint on the exposure-response relationship only over lags 0 to 5, meaning it could be violated at other lags (> 5 in this example). Note that the actual range for slicing depends on the specific basis functions used for the opposite dimension (see the \code{range} argument in \link{shapeConstr} for additional details). When \code{overall = TRUE}, slicing will only constraint the cumulative association over the range specific by \code{slice} (see above).
}

\subsection{Note}{

These methods for \link[dlnm:crossbasis]{crossbasis} objects only allow to specify a single constraint to one dimension of the exposure-lag-response association. However, the user can apply multiple constraints, possibly to both dimensions, by adding multiple terms for the same object in the \link[=buildCmat]{constr} formula (see Examples below).
}

\subsection{Warnings}{

In the current implementation, constraints applied to the lag dimension (when \code{dim = "lag"}) can be straightforwardly defined when the exposure-response function is linear, namely for simpler DLMs. For non-linear exposure-responses, the centering performed by \link[dlnm:crosspred]{crosspred} when predicting the exposure-lag-response relationship can break the constraint definition and return unpredictable results. Users are advised not to apply constraints in the lag dimension of complex DLNMs, or to do so with particular care.
}
}
\examples{
###############################################################################
# Constrained distributed lag models (DLM)

library(dlnm)
library(splines)

#----------------
# Association between PM10 and mortality
#----------------

# Number of years and dow
ny <- length(unique(format(london$date, "\%Y")))

# Create a flexible crossbasis
dlm <- crossbasis(london$pm10, lag = 15, argvar = list(fun = "lin"),
  arglag = list(fun = "ns", knots = logknots(15, df = 4)))

#----- Fit models

# We always have to set `dim = "lag"` because it is set to "var" by default

# Fit with bound constraint, value = 0 by default
bm <- glm(age0_64 ~ dlm + ns(date, df = 7 * ny) + dow, data = london,
  family = "quasipoisson", method = "cirls.fit",
  constr = ~ bound(dlm, dim = "lag"))

# Get a decaying lag-response function with a decreasing constraint
dm <- glm(age0_64 ~ dlm + ns(date, df = 7 * ny) + dow, data = london,
  family = "quasipoisson", method = "cirls.fit",
  constr = ~ shape(dlm, dim = "lag", shape = "dec"))

# Both can be used at the same time
dbm <- glm(age0_64 ~ dlm + ns(date, df = 7 * ny) + dow, data = london,
  family = "quasipoisson", method = "cirls.fit",
  constr = ~ shape(dlm, dim = "lag", shape = "dec") + bound(dlm, dim = "lag"))

#----- Plot to compare

# Compare to unconstrained model
um <- glm(age0_64 ~ dlm + ns(date, df = 7 * ny) + dow, data = london,
  family = "quasipoisson")

# Use crosspred to get the lag-response function
ucp <- crosspred(dlm, um, cen = 0, at = 10)
bcp <- crosspred(dlm, bm, cen = 0, at = 10)
dcp <- crosspred(dlm, dm, cen = 0, at = 10)
dbcp <- crosspred(dlm, dbm, cen = 0, at = 10)

# Now plot
par(mfrow = c(2,2))
plot(ucp, ptype = "slices", lwd = 2, var = 10, main = "Unconstrained")
plot(bcp, ptype = "slices", lwd = 2, var = 10, col = 2,
  main = "Bound")
plot(dcp, ptype = "slices", lwd = 2, var = 10, col = 3,
  main = "Decreasing")
plot(dbcp, ptype = "slices", lwd = 2, var = 10, col = 4,
  main = "Bound & decreasing")


###############################################################################
# Constrained DLNM: var dimension

\dontrun{

library(dlnm)
library(splines)

#----------------
# Association between temperature and mortality
#----------------

# Number of years and dow
ny <- length(unique(format(london$date, "\%Y")))

# Create a flexible crossbasis
cb <- crossbasis(london$tmean, lag = 21,
  argvar = list(fun = "bs", degree = 2, df = 10),
  arglag = list(fun = "ns", knots = logknots(21, df = 5)))

#----- Fit several models

# In the following, `dim` is not shown as it is set to "var" by default

# A nondecreasing model
icm <- glm(age0_64 ~ cb + ns(date, df = 7 * ny) + dow, data = london,
  family = "quasipoisson", method = "cirls.fit",
  constr = ~ shape(cb, shape = "inc"))

# A convex model
ccm <- glm(age0_64 ~ cb + ns(date, df = 7 * ny) + dow, data = london,
  family = "quasipoisson", method = "cirls.fit",
  constr = ~ shape(cb, shape = "cvx"))

# A convex model with a constraint on overall only
ocm <- glm(age0_64 ~ cb + ns(date, df = 7 * ny) + dow, data = london,
  family = "quasipoisson", method = "cirls.fit",
  constr = ~ shape(cb, shape = "cvx", overall = TRUE))

# A convex model effectively constraining only lags 0 and 1
subcm <- glm(age0_64 ~ cb + ns(date, df = 7 * ny) + dow, data = london,
  family = "quasipoisson", method = "cirls.fit",
  constr = ~ shape(cb, shape = "cvx", slice = c(0, 1)))

#----- Plot results

# Plot non-decreasing
icp <- crosspred(cb, icm, cen = 20)
par(mfrow = c(1, 2))
plot(icp, main = "Non-decreasing")
plot(icp, ptype = "overall")

# Plot convex
ccp <- crosspred(cb, ccm, cen = 20)
par(mfrow = c(1, 2))
plot(ccp, main = "Convex")
plot(ccp, ptype = "overall")

# This one is restricted to overall but not for every single lag
ocp <- crosspred(cb, ocm, cen = 20)
par(mfrow = c(1, 2))
plot(ocp, main = "Overall convex")
plot(ocp, ptype = "overall")

# Restricted to lags 0 and 1 only
subcp <- crosspred(cb, subcm, cen = 20)
par(mfrow = c(1, 2))
plot(subcp, main = "Convex")
plot(subcp, ptype = "overall")

}
}
\references{
Gasparrini, A., Armstrong, B., Kenward, M.G., 2010. Distributed lag non-linear models. \emph{Statistics in Medicine} \strong{29}, \strong{2224–2234}. \href{https://doi.org/10.1002/sim.3940}{DOI:10.1002/sim.3940}

Gasparrini, A., Armstrong, B., 2013. Reducing and meta-analysing estimates from distributed lag non-linear models. \emph{BMC Medical Research Methodology} \strong{13}, \strong{1}. \href{https://doi.org/10.1186/1471-2288-13-1}{DOI:10.1186/1471-2288-13-1}
}
\seealso{
\link[dlnm:crossbasis]{crossbasis} for the definition of DLMs and DLNMs as well as their dimensions. \link{shapeConstr} and \link{boundConstr} for generic constraint methods. \link{buildCmat} for how to specify constraints.
}
