% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulCoef.R, R/confint.cirls.R, R/vcov.cirls.R
\name{simulCoef}
\alias{simulCoef}
\alias{confint.cirls}
\alias{vcov.cirls}
\title{Methods for inference on the coefficients of a \code{cirls} object.}
\usage{
simulCoef(object, nsim = 1, seed = NULL, complete = TRUE,
  constrained = TRUE)

\method{confint}{cirls}(object, parm, level = 0.95, nsim = 1000,
  complete = TRUE, ...)

\method{vcov}{cirls}(object, complete = TRUE, nsim = 1000,
  constrained = TRUE, ...)
}
\arguments{
\item{object}{A fitted \code{cirls} object.}

\item{nsim}{The number of simulations to perform.}

\item{seed}{Either \code{NULL} or an integer that will be used in a call to \code{\link[=set.seed]{set.seed()}} before simulating the coefficients.}

\item{complete}{If \code{FALSE}, it does not return inference for undetermined coefficients in case of an over-determined model.}

\item{constrained}{A logical switch indicating whether to simulate from the constrained (the default) or unconstrained (see \link[=uncons]{uncons}) coefficients distribution. When set to \code{FALSE} in \code{vcov}, returns the untruncated covariance matrix as in an unconstrained GLM.}

\item{parm}{A specification of which parameters to compute the confidence intervals for. Either a vector of index numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{The confidence level required.}

\item{...}{Further arguments passed to or from other methods. For \code{vcov} and \code{confint}, it can be used to provide a \code{seed} for the internal coefficient simulation.}
}
\value{
For \code{simulCoef}, a matrix with \code{nsim} rows containing simulated coefficients.

For \code{confint}, a two-column matrix with columns giving lower and upper confidence limits for each parameter.

For \code{vcov}, a matrix of the estimated covariances between the parameter estimates of the model.
}
\description{
\code{simulCoef} simulates coefficients for a fitted \code{cirls} object and uses these simulations for inference. \code{confint} and \code{vcov} directly compute confidence intervals and the variance-covariance matrix for coefficients from a fitted \code{cirls} object. These methods for \code{cirls} objects supersede the default \code{glm} methods.
}
\details{
\subsection{Coefficient inference}{

To perform inference for coefficients the \code{simulCoef} function simulates from the distribution of \eqn{\mathbf{C}\beta} which follows a \strong{Truncated Multivariate Normal} distribution \eqn{TMVN(\mathbf{C}\beta^{*}, \mathbf{C}\mathbf{\Sigma}^{*}\mathbf{C}^{T}, \mathbf{l}, \mathbf{u})} where \eqn{\mathbf{C}} is the constraint matrix with bound vectors \eqn{\mathbf{l}} and \eqn{\mathbf{u}}, and \eqn{\beta^{*}} and \eqn{\mathbf{\Sigma}^{*}} are the unconstrained coefficient vector and variance matrix. The TMVN simulations are then back-transformed to the domain of \eqn{\beta} to allow for inference.
}

\subsection{Functions}{

\code{simulCoef} is the workhorse of the inference and is called internally by \code{confint} and \code{vcov}. All of these are custom methods for \link[=cirls.fit]{cirls} objects that supersede the default methods used for \link[stats:glm]{glm} objects. \code{simulCoef} does not need to be used directly for confidence intervals and variance-covariance matrices, but it can be used to check other summaries of the coefficients distribution.
}
}
\note{
These methods only work when there are less constraints than variables in \code{cirls} model, i.e. when \code{Cmat} has less rows than columns.
}
\examples{
###############################################################################
# Monotone strata levels with the warming dataset

### Fit the model

# Non-decreasing constraint on decadal strata
model <- glm(anomaly ~ decade, data = warming, method = "cirls.fit",
  cons = ~ shape(decade, "inc"))

# Plot result
plot(anomaly ~ year, data = warming, xlab = "", ylab = "Temperature anomaly")
lines(warming$year, predict(model), col = 2, lwd = 2)

### Extract results

# Coefficients and confidence intervals
betas <- coef(model)
v <- vcov(model)
cis <- confint(model)

# Degrees of freedom: represent number of strata change
# ?edf
edf(model)

### Compare with an unconstrained model

# Refit the model without constraints
umodel <- uncons(model)

# Add result
lines(warming$year, predict(umodel), col = 3, lwd = 2, lty = 2)

}
\references{
Geweke, J.F., 1996. Bayesian Inference for Linear Models Subject to Linear Inequality Constraints, in: Lee, J.C., Johnson, W.O., Zellner, A. (Eds.), Modelling and Prediction Honoring Seymour Geisser. \emph{Springer, New York, NY}, pp. 248–263. \href{https://doi.org/10.1007/978-1-4612-2414-3_15}{DOI:10.1007/978-1-4612-2414-3_15}

Botev, Z.I., 2017, The normal law under linear restrictions: simulation and estimation via minimax tilting, \emph{Journal of the Royal Statistical Society, Series B}, \strong{79} (\strong{1}), pp. 1–24. \href{https://doi.org/10.1111/rssb.12162}{DOI:10.1111/rssb.12162}
}
\seealso{
\link[TruncatedNormal:tmvnorm]{rtmvnorm} which is the routine used internally to simulate from a TMVN. \link[=checkCmat]{checkCmat} to check if the constraint matrix can be reduced.
}
