% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulCoef.R, R/confint.cirls.R,
%   R/confint.sim.cirls.R, R/vcov.cirls.R, R/vcov.sim.cirls.R
\name{simulCoef}
\alias{simulCoef}
\alias{confint.cirls}
\alias{confint.sim.cirls}
\alias{vcov.cirls}
\alias{vcov.sim.cirls}
\title{Methods for inference on the coefficients of a \code{cirls} object.}
\usage{
simulCoef(object, nsim = 1, seed = NULL, complete = TRUE,
  constrained = TRUE)

\method{confint}{cirls}(object, parm, level = 0.95, nsim = 1000,
  complete = TRUE, ...)

\method{confint}{sim.cirls}(object, parm, level = 0.95, ...)

\method{vcov}{cirls}(object, complete = TRUE, nsim = 1000,
  constrained = TRUE, ...)

\method{vcov}{sim.cirls}(object, ...)
}
\arguments{
\item{object}{An object of class \code{cirls} or \code{sim.cirls}.}

\item{nsim}{The number of simulations to perform.}

\item{seed}{Either \code{NULL} or an integer that will be used in a call to \code{\link[=set.seed]{set.seed()}} before simulating the coefficients.}

\item{complete}{If \code{FALSE}, it does not return inference for undetermined coefficients in case of an over-determined model.}

\item{constrained}{A logical switch indicating whether to simulate from the constrained (the default) or unconstrained (see \link{uncons}) coefficients distribution. When set to \code{FALSE} in \code{vcov}, returns the untruncated covariance matrix as in an unconstrained GLM.}

\item{parm}{A specification of which parameters to compute the confidence intervals for. Either a vector of index numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{The confidence level required.}

\item{...}{Further arguments passed to or from other methods. For \code{vcov} and \code{confint}, it can be used to provide a \code{seed} for the internal coefficient simulation.}
}
\value{
For \code{simulCoef}: a \code{sim.cirls} object, that is a matrix of \code{nsim} rows containing simulated coefficients with the attributes \code{seed}, \code{complete} and \code{constrained}.

For \code{confint} methods: a two-column matrix with columns giving lower and upper confidence limits for each parameter.

For \code{vcov} methods: a matrix of the estimated covariance between the parameter estimates of the model.
}
\description{
\code{simulCoef} simulates coefficients for a fitted \code{cirls} object and uses these simulations for inference. \code{confint} and \code{vcov} directly compute confidence intervals and the variance-covariance matrix for coefficients from a fitted \code{cirls} object. These methods for \code{cirls} objects supersede the default \code{glm} methods.
}
\details{
\subsection{Coefficient inference}{

To perform inference for coefficients the \code{simulCoef} function simulates from the distribution of \eqn{\mathbf{C}\beta} which follows a \strong{Truncated Multivariate Normal} distribution \eqn{TMVN(\mathbf{C}\beta^{*}, \mathbf{C}\mathbf{\Sigma}^{*}\mathbf{C}^{T}, \mathbf{l}, \mathbf{u})} where \eqn{\mathbf{C}} is the constraint matrix with bound vectors \eqn{\mathbf{l}} and \eqn{\mathbf{u}}, and \eqn{\beta^{*}} and \eqn{\mathbf{\Sigma}^{*}} are the unconstrained coefficient vector and variance matrix. The TMVN simulations are then back-transformed to the domain of \eqn{\beta} to allow for inference.
}

\subsection{Functions}{

\code{simulCoef} is the workhorse of the inference, performing the simulations described above. It is called internally by \code{confint.cirls} and \code{vcov.cirls} to compute confidence intervals and the variance-covariance matrix, respectively. These are custom methods for \link[=cirls.fit]{cirls} objects that supersede the default methods used for \link[stats:glm]{glm} objects. Alternatively \code{simulCoef} can be used directly to simulate \code{nsim} coefficient to then be passed to \code{vcov.sim.cirls} and \code{confint.sim.cirls} methods. This avoids simulating several times when both confidence intervals and variance-covariance are needed for instance.
}
}
\note{
These methods only work when there are less constraints than variables in \code{cirls} model, i.e. when \code{Cmat} has less rows than columns.
}
\examples{
###############################################################################
# Monotone strata levels with the warming dataset

### Typical usage

# Non-decreasing constraint on decadal strata
model <- glm(anomaly ~ decade, data = warming, method = "cirls.fit",
  cons = ~ shape(decade, "inc"))

# Extract covariance matrix and confidence intervals
v <- vcov(model)
ci <- confint(model)

### Using the sim.cirls objects

# Simulate coefficients directly
sims <- simulCoef(model, nsim = 1000, seed = 4)

# Get covariance matrix and confidence intervals
vs <- vcov(sims)
cis <- confint(sims)

# Check that we get the same results as the typical usage
v2 <- vcov(model, seed = 4)
ci2 <- confint(model, seed = 4)
identical(vs, v2)
identical(cis, ci2)
}
\references{
Geweke, J.F., 1996. Bayesian Inference for Linear Models Subject to Linear Inequality Constraints, in: Lee, J.C., Johnson, W.O., Zellner, A. (Eds.), Modelling and Prediction Honoring Seymour Geisser. \emph{Springer, New York, NY}, pp. 248–263. \href{https://doi.org/10.1007/978-1-4612-2414-3_15}{DOI:10.1007/978-1-4612-2414-3_15}

Botev, Z.I., 2017, The normal law under linear restrictions: simulation and estimation via minimax tilting, \emph{Journal of the Royal Statistical Society, Series B}, \strong{79} (\strong{1}), pp. 1–24. \href{https://doi.org/10.1111/rssb.12162}{DOI:10.1111/rssb.12162}
}
\seealso{
\link[TruncatedNormal:tmvnorm]{rtmvnorm} which is the routine used internally to simulate from a TMVN. \link{reduceCons} to reduce the set of constraints.
}
