% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zerosumConstr.R
\name{zerosumConstr}
\alias{zerosumConstr}
\title{Zero-sum constraint matrix}
\usage{
zerosumConstr(..., group = FALSE)
}
\arguments{
\item{...}{Variables to be included in the constraint.}

\item{group}{If set to TRUE, the constraint is build independently for each variable in \code{...}}
}
\description{
Build constraint matrix and bounds for coefficients summing to zero.
}
